api/requirements.txt
Django==4.0.2
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.2
gevent==21.12.0
gunicorn==20.1.0
Pillow==9.0.1
psycopg2==2.9.3
